Move suggestion code to challenge related folders
