+++ packages/server/src/app/challenge/controllers/challenge_suggestion.controller.js
@@ -1,4 +1,4 @@
-import { ChallengeSuggestionService } from './services/challenge_suggestion.service.js';
+import { ChallengeSuggestionService } from '../../challenge_suggestion/services/challenge_suggestion.service.js';
 
 class ChallengeSuggestionController {
   /** @return {Promise<ChallengeSuggestionEntity[]>} */
