+++ packages/server/src/app/challenge/entities/challenge_suggestion.entity.js
@@ -1,6 +1,6 @@
 import { Model } from 'sequelize';
-import definition from '../../db/table-definitions/challenge-suggestion.js';
-import { DatabaseService } from '../core/services/index.js';
+import definition from '../../../db/table-definitions/challenge-suggestion.js';
+import { DatabaseService } from '../../core/services/index.js';
 
 export class ChallengeSuggestionEntity extends Model {}
