+++ packages/server/src/app/challenge/routers/challenge.router.js
@@ -1,11 +1,14 @@
 import express from 'express';
 import { jwtAuthHeaderValidator } from '../../auth/index.js';
 import { userChallengeController } from '../controllers/challenge.controller.js';
+import { challengeSuggestionRouter } from './challenge_suggestion.router.js';
 
-export const userChallengeRouter = express.Router();
+export const challengeRouter = express.Router();
 
-userChallengeRouter.get('/', jwtAuthHeaderValidator, async (request, response) => {
+challengeRouter.get('/', jwtAuthHeaderValidator, async (request, response) => {
   const challenges = await userChallengeController.getForUser(request.userId);
 
   response.send(challenges);
 });
+
+challengeRouter.use('/suggestion', challengeSuggestionRouter);
