+++ packages/server/src/app/challenge_suggestion/services/challenge_suggestion.service.js
@@ -1,4 +1,4 @@
-import { challengeSuggestionRepository } from '../challenge_suggestion.repository.js';
+import { challengeSuggestionRepository } from '../../challenge/repositories/challenge_suggestion.repository.js';
 
 export class ChallengeSuggestionService {
   /** @return {ChallengeSuggestionService} */
